refactor(firestore-bigquery-export): refactor import script
